+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-06-30 22:28:22
-// Commit: test(i18n): add unit tests for i18n business logic
+// Updated at: 2026-07-02 13:07:06
+// Commit: perf(i18n): memoize expensive computations in i18n
 
 class i18nComponentState {
-  final String commitRef = '2440';
-  final String lastMessage = 'test(i18n): add unit tests for i18n business logic';
-  final DateTime timestamp = DateTime.parse('2026-06-30T22:28:22+05:00');
+  final String commitRef = '2466';
+  final String lastMessage = 'perf(i18n): memoize expensive computations in i18n';
+  final DateTime timestamp = DateTime.parse('2026-07-02T13:07:06+05:00');
 }